chore: Update Streamlit configuration
Update the Streamlit configuration in the `.streamlit/config.toml` file. Remove the `port` setting and keep only the `maxUploadSize` setting with a value of 1000. This change is necessary to optimize the server configuration for handling file uploads.

Refactor the Streamlit code in the `main.py` file for the DESD (Detección de Errores en Documentos Escaneados) application. This includes setting the logo, displaying an image, adding a subheader and caption, and creating page links.

Add the `pages/duplicidad.py` file for the DuDe (Duplicate Detection) feature. The code includes setting the logo, displaying a title and caption, uploading files, processing the uploaded images, checking the status of the processing, and providing options to download the results and clean up.

Update the `requirements.txt` file with the necessary dependencies for the Streamlit application, including `streamlit-cookies-controller`, `streamlit`, `aiohttp`, `requests`, `python-dotenv`, `pandas`, and `numpy`.