feat: Update lateral menu logo size and caption
This commit updates the lateral menu in the `menu.py` file. It modifies the size of the logo image to be 100 pixels wide and removes the commented out code for the logo. Additionally, it updates the caption to display the version number and the user's name. These changes improve the visual appearance and provide relevant information to the user.